chore: move auth typeDefs to own file
