+++ packages/api/lib/graphql/types/auth.ts
@@ -1,0 +1,23 @@
+import { gql } from 'apollo-server-express'
+
+export const typeDefs = gql`
+  type AuthPayload {
+    id: ID!
+    token: String!
+    email: String!
+    username: String!
+    language: Languages
+  }
+
+  type LogoutResponse {
+    status: String!
+    message: String!
+  }
+
+  input RegisterInput {
+    email: String! @isEmail
+    username: String!
+    password: String!
+    confirmPassword: String!
+  }
+`
